+++ packages/client/src/features/tools/marquee-selection/model.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2021-2023 EclipseSource and others.
+ * Copyright (c) 2021-2026 EclipseSource and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,9 +14,10 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { boundsFeature, Point, RectangularNode } from '@eclipse-glsp/sprotty';
+import { feedbackFeature } from '../../../base/feedback/feedback-action-dispatcher';
 
 export class MarqueeNode extends RectangularNode {
-    static override readonly DEFAULT_FEATURES = [boundsFeature];
+    static override readonly DEFAULT_FEATURES = [boundsFeature, feedbackFeature];
     startPoint: Point;
     endPoint: Point;
 }